refactor: clarify picker formatter ownership
