+++ src/components/HarnessPicker.tsx
@@ -2,8 +2,8 @@ import { useNavigate } from "react-router";
 import type { HarnessSummary } from "@aws-sdk/client-bedrock-agentcore-control";
 import type { ScreenProps } from "../handlers/types";
 import { coreOptsFromCtx } from "../handlers/utils";
+import { formatTimestamp } from "./formatTimestamp";
 import { PaginatedTablePicker } from "./PaginatedTablePicker";
-import { formatTimestamp } from "./pickerFormatters";
 import type { DataTableColumn } from "./ui/data-table";
 
 // HarnessRow is the flat, display-ready shape the table renders. It also satisfies
