+++ src/components/RuntimeVersionPicker.tsx
@@ -2,8 +2,8 @@ import type { AgentRuntime } from "@aws-sdk/client-bedrock-agentcore-control";
 import { useNavigate } from "react-router";
 import type { ScreenProps } from "../handlers/types";
 import { coreOptsFromCtx } from "../handlers/utils";
+import { formatTimestamp } from "./formatTimestamp";
 import { PaginatedTablePicker } from "./PaginatedTablePicker";
-import { formatTimestamp } from "./pickerFormatters";
 import type { DataTableColumn } from "./ui/data-table";
 
 interface RuntimeVersionRow extends Record<string, unknown> {
