+++ src/components/formatTimestamp.ts
@@ -3,8 +3,3 @@ export function formatTimestamp(value: unknown): string {
   if (Number.isNaN(date.getTime())) return "";
   return date.toISOString().slice(0, 16).replace("T", " ");
 }
-
-export function runtimeIdSuffix(value: unknown): string {
-  const id = String(value ?? "");
-  return id.slice(id.lastIndexOf("-") + 1);
-}
